Update e ligtas

+++ app/Events/IncidentReportEvent.php
@@ -3,6 +3,7 @@
 namespace App\Events;
 
 use App\Models\IncidentReport;
+use Carbon\Carbon;
 use Illuminate\Broadcasting\Channel;
 use Illuminate\Contracts\Broadcasting\ShouldBroadcast;
 use Illuminate\Foundation\Events\Dispatchable;
@@ -12,10 +13,11 @@ class IncidentReportEvent implements ShouldBroadcast
 {
     use Dispatchable, SerializesModels;
 
-    public $reportLog, $incidentReport;
+    public $reportLog, $incidentReport, $totalReport;
 
     public function __construct()
     {
+        $this->totalReport = IncidentReport::where('report_time', '>=', Carbon::now()->format('Y-m-d H:i:s'))->count();
     }
 
     function approveStatus($accidentReportId)
@@ -37,10 +39,13 @@ class IncidentReportEvent implements ShouldBroadcast
     function revertIncidentReport($accidentReportId, $reportPhotoPath)
     {
         $this->incidentReport = new IncidentReport;
-        $image_path           = public_path('reports_image/' . $reportPhotoPath);
 
-        if (file_exists($image_path)) {
-            unlink($image_path);
+        if ($reportPhotoPath) {
+            $image_path = public_path('reports_image/' . $reportPhotoPath);
+
+            if (file_exists($image_path)) {
+                unlink($image_path);
+            }
         }
 
         $this->incidentReport->find($accidentReportId)->delete();
